refactor: searchAddress api 위치 이전 및 리팩토링
